+++ src/main/java/com/prgrms/needit/domain/user/user/dto/LoginRequest.java
@@ -1,4 +1,4 @@
-package com.prgrms.needit.domain.user.login.dto;
+package com.prgrms.needit.domain.user.user.dto;
 
 import javax.validation.constraints.NotBlank;
 import lombok.Builder;
