+++ src/main/java/com/prgrms/needit/domain/user/user/service/UserService.java
@@ -1,4 +1,4 @@
-package com.prgrms.needit.domain.user.login.service;
+package com.prgrms.needit.domain.user.user.service;
 
 import com.prgrms.needit.common.config.jwt.JwtTokenProvider;
 import com.prgrms.needit.common.domain.dto.IsUniqueRequest;
@@ -8,11 +8,11 @@ import com.prgrms.needit.common.error.ErrorCode;
 import com.prgrms.needit.common.error.exception.NotFoundResourceException;
 import com.prgrms.needit.domain.user.center.entity.Center;
 import com.prgrms.needit.domain.user.center.repository.CenterRepository;
-import com.prgrms.needit.domain.user.login.dto.CurUser;
-import com.prgrms.needit.domain.user.login.dto.LoginRequest;
-import com.prgrms.needit.domain.user.login.dto.TokenResponse;
 import com.prgrms.needit.domain.user.member.entity.Member;
 import com.prgrms.needit.domain.user.member.repository.MemberRepository;
+import com.prgrms.needit.domain.user.user.dto.CurUser;
+import com.prgrms.needit.domain.user.user.dto.LoginRequest;
+import com.prgrms.needit.domain.user.user.dto.TokenResponse;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
